refactor: move the amd64 instruction encoder to a shared src/encoder/amd64
