+++ src/core/compiler/backend/railshot/fp.go
@@ -15,7 +15,7 @@ func floatBits(v float64, f64 bool) uint64 {
 }
 
 // Floating point (f32/f64), ported from WARP's SSE lowering with the NaN- and
-// signed-zero-correct sequences backend/railshot/amd64 uses. Floats are handled eagerly:
+// signed-zero-correct sequences src/encoder/amd64 uses. Floats are handled eagerly:
 // operands are materialized into XMM registers by a parallel allocator (fregUser/
 // fpinned) and the result is pushed as an XMM-resident value. XMM and GP register
 // namespaces are independent, so the integer condense engine is untouched.
@@ -353,7 +353,7 @@ func (f *fn) i2fU(f64, srcWide bool) {
 
 // truncLimitBits returns the exclusive source-width float bounds outside which a
 // trunc to the given integer type must trap (x valid iff min < x < max). Mirrors
-// backend/railshot/amd64 / WARP FloatTruncLimitsExcl.
+// src/encoder/amd64 / WARP FloatTruncLimitsExcl.
 func truncLimitBits(signed, f64src, dstWide bool) (minBits, maxBits uint64) {
 	switch {
 	case !f64src && signed && !dstWide:
