+++ src/encoder/amd64/asm_sse.go
